backend/requirements.txt
asgiref==3.5.2
autopep8==2.0.1
Django==4.1.2
django-cors-headers==3.13.0
djangorestframework==3.13.1
Pillow==9.3.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
python-dotenv==0.21.1
pytz==2022.2.1
sqlparse==0.4.2
tomli==2.0.1
